requirements.txt
# Core Web App
streamlit
pillow

# Data Handling
pandas
numpy

# Visualization (optional, for plots and performance display)
matplotlib
seaborn

# YOLOv5 + Ultralytics
torch
torchvision
ultralytics

# OpenCV for image processing (used in YOLOv5)
opencv-python

# Optional: for tqdm bars/logs
tqdm
